Extract asset management routes into blueprints/asset_routes.py
Phase 3 of breaking up the quickstart.py monolith. Moved the 12 routes
that manage library images, custom fonts, and overlay previews into a
new blueprint, plus the shared globals/helpers they (and the /step
renderer) depend on into modules/assets.py.

list_overlay_fonts() and build_preview_image_data() are also called
from the /step renderer and inject_version_info's context processor,
which stays in quickstart.py — both are re-imported under their
original names so those call sites are untouched.

_safe_join/_safe_rel_path had no quickstart.py-specific state (pure
path-containment checks) and are also used by unrelated code
(config rename/copy), so they moved to modules/helpers.py as
safe_join/safe_rel_path with the usual backward-compatible aliases.

Found and fixed a real bug while moving serve_preview_image: it had a
try/except block of dead code after its return statement, left over
from an earlier version of the function — unreachable, deleted.

Net: quickstart.py shrinks by 635 lines.
