+++ quickstart.py
@@ -24,11 +24,9 @@ from io import BytesIO
 from threading import Thread
 from pathlib import Path
 from collections import deque
-from urllib.parse import urlparse
 
 import namesgenerator
 import requests
-from PIL import Image, ImageDraw, ImageFont, ImageColor
 from cachelib.file import FileSystemCache
 from datetime import datetime, timezone, timedelta
 from dotenv import load_dotenv
@@ -42,7 +40,6 @@ from flask import (
     flash,
     session,
     send_file,
-    send_from_directory,
     abort,
     has_request_context,
 )
@@ -67,11 +64,15 @@ from modules.background_jobs import (
     fail_background_job as _fail_background_job,
 )
 from blueprints.validation_routes import bp as validation_routes_bp, refresh_plex_libraries
+from blueprints.asset_routes import bp as asset_routes_bp
+from modules.assets import build_preview_image_data as _build_preview_image_data, list_overlay_fonts
 
 Request.max_form_parts = 100000  # Allow more form fields if needed
 
 _resolve_request_config_name = persistence.resolve_request_config_name
 utc_now_iso = helpers.utc_now_iso
+_safe_join = helpers.safe_join
+_safe_rel_path = helpers.safe_rel_path
 
 ACTIVE_WORK_POLICIES = {
     "kometa_run": [
@@ -3915,45 +3916,6 @@ def _normalize_config_filename(config_name: str | None) -> str:
     return name or "default"
 
 
-def _safe_rel_path(raw_path: str | None, allow_subdirs: bool = False) -> str | None:
-    if not isinstance(raw_path, str):
-        return None
-    raw_path = raw_path.strip()
-    if not raw_path:
-        return None
-    if "\x00" in raw_path:
-        return None
-
-    drive, _ = os.path.splitdrive(raw_path)
-    if drive:
-        return None
-    if os.path.isabs(raw_path):
-        return None
-
-    normalized = os.path.normpath(raw_path)
-    if normalized in (".", ""):
-        return None
-    if normalized.startswith("..") or normalized.startswith("../") or normalized.startswith("..\\"):
-        return None
-    if not allow_subdirs and ("/" in normalized or "\\" in normalized):
-        return None
-
-    return normalized
-
-
-def _safe_join(base_dir: str | Path, raw_path: str | None, allow_subdirs: bool = False) -> Path | None:
-    rel = _safe_rel_path(raw_path, allow_subdirs=allow_subdirs)
-    if not rel:
-        return None
-    try:
-        base = Path(base_dir).resolve()
-        candidate = (base / rel).resolve()
-        candidate.relative_to(base)
-        return candidate
-    except Exception:
-        return None
-
-
 def _resolve_user_dir(raw_path: str | None) -> Path | None:
     if not isinstance(raw_path, str):
         return None
@@ -4095,104 +4057,6 @@ def _rename_config_files(old_name: str, new_name: str, dry_run: bool = False) ->
 DOTENV = os.path.relpath(os.path.join(helpers.CONFIG_DIR, ".env"))
 load_dotenv(DOTENV, override=True)
 
-UPLOAD_FOLDER = os.path.join(helpers.CONFIG_DIR, "uploads")
-UPLOAD_FOLDERS = {
-    "movie": os.path.join(UPLOAD_FOLDER, "movies"),
-    "show": os.path.join(UPLOAD_FOLDER, "shows"),
-    "season": os.path.join(UPLOAD_FOLDER, "seasons"),
-    "episode": os.path.join(UPLOAD_FOLDER, "episodes"),
-}
-# Ensure all upload subdirectories exist
-for folder in UPLOAD_FOLDERS.values():
-    os.makedirs(folder, exist_ok=True)
-IMAGES_FOLDER = os.path.join(helpers.MEIPASS_DIR, "static", "images")
-OVERLAY_FOLDER = os.path.join(IMAGES_FOLDER, "overlays")
-FONTS_FOLDER = os.path.join(helpers.MEIPASS_DIR, "static", "fonts")
-DEFAULT_IMAGE_MAP = {
-    "movie": os.path.join(IMAGES_FOLDER, "default.png"),
-    "show": os.path.join(IMAGES_FOLDER, "default-sho_preview.png"),
-    "season": os.path.join(IMAGES_FOLDER, "default-season_preview.png"),
-    "episode": os.path.join(IMAGES_FOLDER, "default-episode_preview.png"),
-}
-BUILTIN_PREVIEW_IMAGES = (
-    "overlay_alignment_guide.png",
-    "overlay_alignment_guide_episodes.png",
-)
-BUILTIN_PREVIEW_IMAGES_BY_TYPE = {
-    "movie": ("overlay_alignment_guide.png",),
-    "show": ("overlay_alignment_guide.png",),
-    "season": ("overlay_alignment_guide.png",),
-    "episode": ("overlay_alignment_guide_episodes.png",),
-}
-PREVIEW_FOLDER = os.path.join(helpers.CONFIG_DIR, "previews")
-os.makedirs(PREVIEW_FOLDER, exist_ok=True)
-OVERLAY_CACHE_FOLDER = os.path.join(helpers.CONFIG_DIR, "cache", "overlays")
-os.makedirs(OVERLAY_CACHE_FOLDER, exist_ok=True)
-OVERLAY_CACHE_TTL_SECONDS = 60 * 60 * 24 * 30
-_FONT_CACHE: dict[str, list[str]] = {}
-
-
-def _list_preview_images_for_type(image_type: str) -> list[str]:
-    """Return built-in guide images plus uploaded images for a preview image type."""
-    builtin_candidates = BUILTIN_PREVIEW_IMAGES_BY_TYPE.get(image_type, ())
-    builtins = [name for name in builtin_candidates if os.path.exists(os.path.join(IMAGES_FOLDER, name))]
-    uploads_dir = UPLOAD_FOLDERS.get(image_type)
-    uploads: list[str] = []
-    if uploads_dir and os.path.exists(uploads_dir):
-        uploads = sorted(
-            [img for img in os.listdir(uploads_dir) if any(img.lower().endswith(f".{ext}") for ext in helpers.ALLOWED_EXTENSIONS)],
-            key=str.casefold,
-        )
-    return builtins + [img for img in uploads if img not in builtins]
-
-
-def _build_preview_image_data() -> dict[str, list[str]]:
-    return {img_type: _list_preview_images_for_type(img_type) for img_type in UPLOAD_FOLDERS}
-
-
-def _resolve_preview_base_image_path(img_type: str, selected_image: str) -> str:
-    if not selected_image or selected_image == "default":
-        return DEFAULT_IMAGE_MAP.get(img_type, DEFAULT_IMAGE_MAP["movie"])
-
-    if selected_image in BUILTIN_PREVIEW_IMAGES and selected_image not in BUILTIN_PREVIEW_IMAGES_BY_TYPE.get(img_type, ()):
-        return DEFAULT_IMAGE_MAP.get(img_type, DEFAULT_IMAGE_MAP["movie"])
-
-    static_candidate = _safe_join(IMAGES_FOLDER, selected_image)
-    if static_candidate and static_candidate.exists():
-        return str(static_candidate)
-
-    upload_candidate = _safe_join(UPLOAD_FOLDERS[img_type], selected_image)
-    if upload_candidate and upload_candidate.exists():
-        return str(upload_candidate)
-
-    return DEFAULT_IMAGE_MAP.get(img_type, DEFAULT_IMAGE_MAP["movie"])
-
-
-# Font discovery (TTF/OTF) across common static dirs
-def list_overlay_fonts() -> list[str]:
-    global _FONT_CACHE
-    config_name = session.get("config_name") if has_request_context() else None
-    if config_name:
-        migration = helpers.migrate_legacy_custom_fonts_to_config(config_name)
-        if migration.get("copied"):
-            _FONT_CACHE = {}
-    cache_key = helpers.normalize_config_name_for_storage(config_name) if config_name else "__default__"
-    cached = _FONT_CACHE.get(cache_key)
-    if cached:
-        return cached
-    fonts: list[str] = []
-    font_dirs = helpers.get_font_dirs(include_static=True, include_custom=True, config_name=config_name)
-    for fdir in font_dirs:
-        try:
-            if os.path.isdir(fdir):
-                for fname in os.listdir(fdir):
-                    if fname.lower().endswith((".ttf", ".otf")) and fname not in fonts:
-                        fonts.append(fname)
-        except Exception:
-            continue
-    _FONT_CACHE[cache_key] = fonts
-    return fonts
-
 
 def _config_name_from_yaml_filename(filename: str | None) -> str | None:
     text = str(filename or "").strip()
@@ -4219,6 +4083,7 @@ kometa_process = None
 app = Flask(__name__)
 
 app.register_blueprint(validation_routes_bp)
+app.register_blueprint(asset_routes_bp)
 
 # Run version check at startup
 app.config["VERSION_CHECK"] = helpers.check_for_update()
@@ -4973,540 +4838,6 @@ app.config["QS_PORT"] = running_port
 debug_mode = args.debug if args.debug else helpers.booler(os.getenv("QS_DEBUG", "0"))
 
 helpers.ts_log(f"Running on port: {port} | Debug Mode: {'Enabled' if debug_mode else 'Disabled'}", level="INFO")
-
-
-@app.route("/upload_library_image", methods=["POST"])
-def upload_library_image():
-    if "image" not in request.files:
-        return jsonify({"status": "error", "message": "No image uploaded"}), 400
-    image = request.files["image"]
-    image_type = request.form.get("type")
-
-    if not image or image_type not in UPLOAD_FOLDERS:
-        return (
-            jsonify({"status": "error", "message": "Invalid request parameters"}),
-            400,
-        )
-
-    # Validate extension
-    filename = secure_filename(image.filename)
-    ext = filename.rsplit(".", 1)[-1].lower()
-    if ext not in helpers.ALLOWED_EXTENSIONS:
-        return (
-            jsonify(
-                {
-                    "status": "error",
-                    "message": f"Invalid file type. Allowed: {helpers.allowed_extensions_string()}",
-                }
-            ),
-            400,
-        )
-
-    # Open and validate image
-    img = Image.open(image)
-    aspect_ratio = "16:9" if image_type == "episode" else "2:3"
-    if not helpers.is_valid_aspect_ratio(img, target_ratio=aspect_ratio):
-        message = "Image must have a 16:9 aspect ratio (e.g., 1920x1080)." if image_type == "episode" else "Image must have a 1:1.5 aspect ratio (e.g., 1000x1500)."
-        return jsonify({"status": "error", "message": message}), 400
-
-    # Resize to target size
-    target_size = (1920, 1080) if image_type == "episode" else (1000, 1500)
-    img = img.resize(target_size, Image.LANCZOS)
-
-    # Save image
-    save_folder = UPLOAD_FOLDERS[image_type]
-    os.makedirs(save_folder, exist_ok=True)
-    save_path = os.path.join(save_folder, filename)
-    base, ext = os.path.splitext(filename)
-    counter = 1
-    while os.path.exists(save_path):
-        filename = f"{base}_{counter}{ext}"
-        save_path = os.path.join(save_folder, filename)
-        counter += 1
-    img.save(save_path)
-
-    return jsonify(
-        {
-            "status": "success",
-            "message": f"Image uploaded and saved as {filename}",
-            "filename": filename,
-        }
-    )
-
-
-@app.route("/upload-fonts", methods=["POST"])
-def upload_fonts():
-    files = request.files.getlist("fonts")
-    if not files:
-        return jsonify({"status": "error", "message": "No fonts uploaded"}), 400
-
-    config_name = (request.form.get("config_name") or session.get("config_name") or "").strip()
-    fonts_dir = helpers.get_custom_fonts_dir(config_name or None)
-
-    os.makedirs(fonts_dir, exist_ok=True)
-    saved = []
-    errors = []
-
-    for font_file in files:
-        if not font_file or not font_file.filename:
-            continue
-        filename = secure_filename(font_file.filename)
-        ext = os.path.splitext(filename)[1].lower()
-        if ext not in helpers.FONT_EXTENSIONS:
-            errors.append(f"Invalid font type: {filename}")
-            continue
-        save_path = os.path.join(str(fonts_dir), filename)
-        font_file.save(save_path)
-        saved.append(filename)
-
-    if saved:
-        global _FONT_CACHE
-        _FONT_CACHE = {}
-
-    if not saved:
-        return jsonify({"status": "error", "message": "No valid fonts uploaded.", "errors": errors}), 400
-
-    return jsonify(
-        {
-            "status": "success",
-            "message": f"Uploaded {len(saved)} font(s).",
-            "saved": saved,
-            "errors": errors,
-            "fonts": list_overlay_fonts(),
-        }
-    )
-
-
-@app.route("/custom-fonts/<path:filename>", methods=["GET"])
-def custom_fonts(filename):
-    safe_name = os.path.basename(filename)
-    if not safe_name or safe_name != filename:
-        abort(404)
-    if not safe_name.lower().endswith((".ttf", ".otf")):
-        abort(404)
-
-    current_config = session.get("config_name") if has_request_context() else None
-    for fdir in helpers.get_font_dirs(include_static=True, include_custom=True, config_name=current_config):
-        candidate = os.path.join(str(fdir), safe_name)
-        if os.path.exists(candidate):
-            return send_from_directory(str(fdir), safe_name)
-
-    abort(404)
-
-
-@app.route("/fetch_library_image", methods=["POST"])
-def fetch_library_image():
-    data = request.json
-    image_url = data.get("url")
-    image_type = data.get("type")
-
-    if not image_url or image_type not in UPLOAD_FOLDERS:
-        return (
-            jsonify({"status": "error", "message": "Invalid request parameters"}),
-            400,
-        )
-
-    valid_url, url_message = url_validation.validate_url(image_url, allow_local=False)
-    if not valid_url:
-        return (
-            jsonify({"status": "error", "message": f"Invalid image URL: {url_message}"}),
-            400,
-        )
-
-    try:
-        response = requests.get(image_url, stream=True, timeout=5)
-        response.raise_for_status()
-        img = Image.open(BytesIO(response.content))
-
-        file_extension = img.format.lower()
-        if file_extension not in helpers.ALLOWED_EXTENSIONS:
-            return (
-                jsonify(
-                    {
-                        "status": "error",
-                        "message": f"Invalid file type. Allowed: {helpers.allowed_extensions_string()}",
-                    }
-                ),
-                400,
-            )
-
-        # Validate aspect ratio
-        aspect_ratio = "16:9" if image_type == "episode" else "2:3"
-        if not helpers.is_valid_aspect_ratio(img, target_ratio=aspect_ratio):
-            message = "Image must have a 16:9 aspect ratio (e.g., 1920x1080)." if image_type == "episode" else "Image must have a 1:1.5 aspect ratio (e.g., 1000x1500)."
-            return jsonify({"status": "error", "message": message}), 400
-
-        # Resize to target size
-        target_size = (1920, 1080) if image_type == "episode" else (1000, 1500)
-        img = img.resize(target_size, Image.LANCZOS)
-
-        # Generate filename
-        filename = secure_filename(os.path.basename(image_url))
-        if "." not in filename or filename.split(".")[-1].lower() not in helpers.ALLOWED_EXTENSIONS:
-            filename += ".png"
-
-        # Save image
-        save_folder = UPLOAD_FOLDERS[image_type]
-        os.makedirs(save_folder, exist_ok=True)
-        save_path = os.path.join(save_folder, filename)
-        base, ext = os.path.splitext(filename)
-        counter = 1
-        while os.path.exists(save_path):
-            filename = f"{base}_{counter}{ext}"
-            save_path = os.path.join(save_folder, filename)
-            counter += 1
-        img.save(save_path)
-
-        return jsonify(
-            {
-                "status": "success",
-                "message": f"Image fetched and saved as {filename}",
-                "filename": filename,
-            }
-        )
-
-    except requests.exceptions.RequestException as e:
-        return (
-            jsonify({"status": "error", "message": f"Failed to fetch image: {str(e)}"}),
-            400,
-        )
-    except Exception as e:
-        return (
-            jsonify({"status": "error", "message": f"Processing error: {str(e)}"}),
-            400,
-        )
-
-
-@app.route("/rename_library_image", methods=["POST"])
-def rename_library_image():
-    data = request.json
-    old_name = data.get("old_name")
-    new_name = data.get("new_name")
-    image_type = data.get("type")
-
-    if not old_name or not new_name or image_type not in UPLOAD_FOLDERS:
-        return jsonify({"status": "error", "message": "Invalid parameters"}), 400
-
-    save_folder = UPLOAD_FOLDERS[image_type]
-    old_path = _safe_join(save_folder, old_name)
-    if not old_path:
-        return jsonify({"status": "error", "message": "Invalid file name"}), 400
-
-    if not old_path.exists():
-        return jsonify({"status": "error", "message": "File not found"}), 404
-
-    old_ext = old_path.suffix
-    safe_new_name = str(new_name).strip()
-    if not safe_new_name:
-        return jsonify({"status": "error", "message": "Invalid parameters"}), 400
-    if old_ext:
-        if "." not in safe_new_name:
-            safe_new_name += old_ext
-        elif not safe_new_name.endswith(old_ext):
-            safe_new_name += old_ext
-
-    new_path = _safe_join(save_folder, safe_new_name)
-    if not new_path:
-        return jsonify({"status": "error", "message": "Invalid file name"}), 400
-    if new_path.exists():
-        return (
-            jsonify({"status": "error", "message": "File with new name already exists"}),
-            400,
-        )
-
-    try:
-        os.rename(old_path, new_path)
-        return jsonify({"status": "success", "message": "File renamed successfully"})
-    except Exception as e:
-        return jsonify({"status": "error", "message": str(e)}), 500
-
-
-@app.route("/list_uploaded_images", methods=["GET"])
-def list_uploaded_images():
-    image_type = request.args.get("type")
-    if image_type not in UPLOAD_FOLDERS:
-        return jsonify({"status": "error", "message": "Invalid image type"}), 400
-
-    return jsonify({"status": "success", "images": _list_preview_images_for_type(image_type)})
-
-
-@app.route("/generate_preview", methods=["POST"])
-def generate_preview():
-    data = request.json
-    img_type = data.get("type", "movie")
-    selected_image = data.get("selected_image", "default.png")
-    library_id = data.get("library_id", "default-library")
-
-    # Lazy-load overlay metadata so we can honor JSON-defined URLs (e.g., edition overlays)
-    if not hasattr(generate_preview, "_overlay_meta"):
-        overlay_cfg = helpers.load_quickstart_overlay_config() or []
-        meta = {}
-        for group in overlay_cfg:
-            for ov in group.get("overlays", []):
-                ov_id = ov.get("id")
-                if ov_id:
-                    meta[ov_id] = ov
-        generate_preview._overlay_meta = meta
-    overlay_meta = getattr(generate_preview, "_overlay_meta", {})
-
-    def fetch_image_from_url(url: str) -> Image.Image | None:
-        try:
-            if not url:
-                return None
-            cache_path = None
-            try:
-                ext = os.path.splitext(urlparse(url).path)[1].lower()
-                if ext not in [".png", ".jpg", ".jpeg", ".webp"]:
-                    ext = ".png"
-                cache_key = hashlib.sha256(url.encode("utf-8")).hexdigest()
-                cache_path = os.path.join(OVERLAY_CACHE_FOLDER, f"{cache_key}{ext}")
-                if os.path.exists(cache_path):
-                    age = time.time() - os.path.getmtime(cache_path)
-                    if age <= OVERLAY_CACHE_TTL_SECONDS:
-                        with Image.open(cache_path) as cached_img:
-                            return cached_img.copy()
-            except Exception:
-                cache_path = None
-
-            resp = requests.get(url, timeout=10)
-            resp.raise_for_status()
-            content = resp.content
-            if cache_path:
-                try:
-                    with open(cache_path, "wb") as handle:
-                        handle.write(content)
-                except Exception as e:
-                    helpers.ts_log(f"Failed to cache overlay image {cache_path}: {e}", level="WARNING")
-            return Image.open(BytesIO(content))
-        except Exception as e:
-            helpers.ts_log(f"Failed to fetch overlay image from {url}: {e}", level="WARNING")
-            return None
-
-    # Normalize overlays from dict (by type) or flat list
-    raw_overlays = data.get("overlays", {})
-    if isinstance(raw_overlays, dict):
-        overlays = raw_overlays.get(img_type, [])
-    elif isinstance(raw_overlays, list):
-        overlays = raw_overlays
-    else:
-        overlays = []
-
-    if img_type not in ["movie", "show", "season", "episode"]:
-        return jsonify({"status": "error", "message": "Invalid type"}), 400
-
-    if not os.path.exists(PREVIEW_FOLDER):
-        os.makedirs(PREVIEW_FOLDER)
-
-    preview_filename = f"{library_id}-{img_type}_preview.png"
-    preview_filepath = os.path.join(PREVIEW_FOLDER, preview_filename)
-
-    # Resolve base image
-    base_image_path = _resolve_preview_base_image_path(img_type, selected_image)
-    if not os.path.exists(base_image_path):
-        fallback_size = (1920, 1080) if img_type == "episode" else (1000, 1500)
-        base_img = Image.new("RGBA", fallback_size, (128, 128, 128, 255))
-        base_img.save(base_image_path)
-
-    if not os.path.exists(base_image_path):
-        return jsonify({"status": "error", "message": "Selected image not found."}), 400
-
-    # Open and resize base image
-    base_img = Image.open(base_image_path).convert("RGBA")
-    size = (1920, 1080) if img_type == "episode" else (1000, 1500)
-    base_img = base_img.resize(size, Image.LANCZOS)
-
-    # Determine filename prefix
-    if img_type == "movie":
-        prefix = "mov-"
-    elif img_type == "episode":
-        prefix = "epi-sho-"
-    elif img_type == "season":
-        prefix = "sho-season-"
-    elif img_type == "show":
-        prefix = "sho-"
-    else:
-        prefix = ""
-
-    def render_runtime_overlay(tv: dict, canvas_size: tuple[int, int]) -> Image.Image | None:
-        try:
-            width, height = canvas_size
-            img = Image.new("RGBA", (width, height), (0, 0, 0, 0))
-            draw = ImageDraw.Draw(img)
-
-            prefix = str(tv.get("text", "Runtime: "))
-            fmt = str(tv.get("format", "<<runtimeH>>h <<runtimeM>>m"))
-            runtime_minutes = tv.get("runtime_minutes", 93)
-            try:
-                runtime_minutes = int(runtime_minutes)
-            except Exception:
-                runtime_minutes = 93
-            runtime_h = runtime_minutes // 60
-            runtime_m = runtime_minutes % 60
-
-            rendered_fmt = (
-                fmt.replace("<<runtimeH>>", str(runtime_h))
-                .replace("<<runtimeM>>", str(runtime_m))
-                .replace("<<runtime_total>>", str(runtime_minutes))
-                .replace("<<runtime>>", str(runtime_minutes))
-            )
-            text = f"{prefix}{rendered_fmt}"
-
-            font_size = tv.get("font_size", 55)
-            try:
-                font_size = int(font_size)
-            except Exception:
-                font_size = 55
-            font_path = str(tv.get("font", "") or "").strip()
-
-            # Resolve font path against known font directory if a basename is given
-            font = None
-            font_candidates = []
-            if font_path:
-                font_candidates.append(font_path)
-                base_font = os.path.basename(font_path)
-                for fdir in helpers.get_font_dirs(include_static=True, include_custom=True):
-                    font_candidates.append(os.path.join(str(fdir), base_font))
-            seen_candidates = set()
-            font_candidates = [c for c in font_candidates if c and not (c in seen_candidates or seen_candidates.add(c))]
-            for candidate in font_candidates:
-                if candidate and os.path.exists(candidate):
-                    try:
-                        font = ImageFont.truetype(candidate, font_size)
-                        break
-                    except Exception:
-                        font = None
-            if font is None:
-                font = ImageFont.load_default()
-
-            color_val = tv.get("font_color", "#FFFFFF")
-            try:
-                fill = ImageColor.getcolor(str(color_val), "RGBA")
-            except Exception:
-                fill = (255, 255, 255, 255)
-
-            margin = 20
-            draw.text((width - margin, height - margin), text, fill=fill, font=font, anchor="rb")
-            return img
-        except Exception as e:
-            helpers.ts_log(f"Failed to render runtime overlay: {e}", level="WARNING")
-            return None
-
-    # Apply overlays with template_variables support
-    for overlay_entry in overlays:
-        if isinstance(overlay_entry, str):
-            overlay_id = overlay_entry
-            template_vars = {}
-        elif isinstance(overlay_entry, dict):
-            overlay_id = overlay_entry.get("id")
-            template_vars = overlay_entry.get("template_variables", {})
-
-            # Normalize booleans to lowercase strings (e.g., True → "true")
-            template_vars = {k: str(v).lower() if isinstance(v, bool) else v for k, v in template_vars.items()}
-        else:
-            continue  # skip invalid overlay data
-
-        # Build filename suffix from all template_variables (sorted for consistency)
-        suffix_parts = [f"{key}_{value}" for key, value in sorted(template_vars.items()) if key in {"style", "size", "color"}]
-        suffix = "_" + "_".join(suffix_parts) if suffix_parts else ""
-        filename = f"{prefix}{img_type}-{overlay_id}{suffix}.png"
-        overlay_path = os.path.join(OVERLAY_FOLDER, filename)
-
-        # Fallback to default overlay if specific style not found
-        if not os.path.exists(overlay_path) and suffix:
-            fallback_filename = f"{prefix}{img_type}-{overlay_id}.png"
-            fallback_path = os.path.join(OVERLAY_FOLDER, fallback_filename)
-            if os.path.exists(fallback_path):
-                overlay_path = fallback_path
-
-        if os.path.exists(overlay_path):
-            if overlay_id == "overlay_runtimes":
-                runtime_img = render_runtime_overlay(template_vars, base_img.size)
-                if runtime_img:
-                    base_img.paste(runtime_img, (0, 0), runtime_img)
-                    continue  # skip default image paste
-
-            overlay_img = Image.open(overlay_path).convert("RGBA")
-            base_img.paste(overlay_img, (0, 0), overlay_img)
-
-            # Stack edition overlay below resolution when enabled.
-            if overlay_id == "overlay_resolution":
-                use_edition = str(template_vars.get("use_edition", "false")).lower() == "true"
-                if use_edition:
-                    bbox = overlay_img.getbbox()
-                    if bbox:
-                        edition_url = overlay_meta.get("overlay_resolution", {}).get("edition_overlay_url")
-                        edition_img = fetch_image_from_url(edition_url) if edition_url else None
-                        if edition_img:
-                            edition_img = edition_img.convert("RGBA")
-                            x_offset = bbox[0]
-                            spacing = 15
-                            y_offset = bbox[3] + spacing
-                            base_img.paste(edition_img, (x_offset, y_offset), edition_img)
-
-    base_img.save(preview_filepath)
-
-    return jsonify({"status": "success", "preview_url": f"/{preview_filepath}"})
-
-
-@app.route("/config/previews/<path:filename>")
-def serve_previews(filename):
-    return send_from_directory(PREVIEW_FOLDER, filename)
-
-
-@app.route("/config/uploads/<path:filename>")
-def serve_uploaded_file(filename):
-    return send_from_directory(UPLOAD_FOLDER, filename)
-
-
-@app.route("/get_preview_image/<img_type>", methods=["GET"])
-def get_preview_image(img_type):
-    preview_filename = f"{img_type}_preview.png"
-    preview_path = os.path.join(PREVIEW_FOLDER, preview_filename)
-
-    if not os.path.exists(preview_path):
-        generate_preview()
-
-    if os.path.exists(preview_path):
-        return send_file(preview_path, mimetype="image/png")
-
-    return jsonify({"status": "error", "message": "Preview image not found"}), 400
-
-
-@app.route("/config/previews/<filename>")
-def serve_preview_image(filename):
-    safe_path = _safe_join(PREVIEW_FOLDER, filename)
-    if safe_path and safe_path.exists():
-        return send_file(safe_path, mimetype="image/png")
-    return send_file(os.path.join(IMAGES_FOLDER, "default.png"), mimetype="image/png")
-    try:
-        data = request.get_json()
-        helpers.ts_log("Received data: %s", data, level="INFO")  # Log the received data
-        return jsonify({"status": "success"})
-    except Exception as e:
-        helpers.ts_log("Error updating libraries: %s", str(e), level="ERROR")
-        return jsonify({"status": "error", "message": str(e)}), 500
-
-
-@app.route("/delete_library_image/<filename>", methods=["DELETE"])
-def delete_library_image(filename):
-    image_type = request.args.get("type")
-
-    if image_type not in UPLOAD_FOLDERS:
-        return jsonify({"status": "error", "message": "Invalid image type"}), 400
-
-    uploads_dir = UPLOAD_FOLDERS[image_type]
-    file_path = _safe_join(uploads_dir, filename)
-    if not file_path:
-        return jsonify({"status": "error", "message": "Invalid file name"}), 400
-
-    if not file_path.exists():
-        return jsonify({"status": "error", "message": "File not found"}), 404
-
-    try:
-        os.remove(file_path)
-        return jsonify({"status": "success", "message": f"Deleted {filename}"}), 200
-    except Exception as e:
-        return jsonify({"status": "error", "message": str(e)}), 500
 
 
 @app.route("/")
